backend/gateway/requirements.txt
anyio==3.6.2
certifi==2023.5.7
cffi==1.15.1
click==8.1.3
cryptography==40.0.2
ecdsa==0.18.0
fastapi==0.95.0
h11==0.14.0
httpcore==0.17.0
httpx==0.24.0
idna==3.4
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.7
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
typing_extensions==4.5.0
uvicorn==0.21.1
